splitted const into multiple files
