+++ frontend/src/components/jobs/result/JobIsRunningAlert.jsx
@@ -6,7 +6,7 @@ import { MdPauseCircleOutline } from "react-icons/md";
 import { IconAlert, IconButton } from "@certego/certego-ui";
 
 import { killJob } from "./jobApi";
-import { jobStatuses } from "../../../constants/constants";
+import { jobStatuses } from "../../../constants/jobConst";
 
 import {
   reportedPluginNumber,
