+++ frontend/src/components/scan/utils/RecentScans.jsx
@@ -5,7 +5,7 @@ import { useNavigate } from "react-router-dom";
 import { Card, CardHeader, CardBody } from "reactstrap";
 import { DateHoverable, Loader } from "@certego/certego-ui";
 import useRecentScansStore from "../../../stores/useRecentScansStore";
-import { jobResultSection } from "../../../constants/constants";
+import { jobResultSection } from "../../../constants/miscConst";
 
 function RecentScansCard({
   pk,
