+++ frontend/src/constants/jobConst.js
@@ -1,7 +1,33 @@
+// IMPORTANT - do not change the order of status
+export const jobStatuses = Object.freeze({
+  PENDING: "pending",
+  RUNNING: "running",
+  ANALYZERS_RUNNING: "analyzers_running",
+  ANALYZERS_COMPLETED: "analyzers_completed",
+  CONNECTORS_RUNNING: "connectors_running",
+  CONNECTORS_COMPLETED: "connectors_completed",
+  PIVOTS_RUNNING: "pivots_running",
+  PIVOTS_COMPLETED: "pivots_completed",
+  VISUALIZERS_RUNNING: "visualizers_running",
+  VISUALIZERS_COMPLETED: "visualizers_completed",
+  REPORTED_WITH_FAILS: "reported_with_fails",
+  REPORTED_WITHOUT_FAILS: "reported_without_fails",
+  KILLED: "killed",
+  FAILED: "failed",
+});
+
+export const jobFinalStatuses = Object.freeze({
+  REPORTED_WITH_FAILS: jobStatuses.REPORTED_WITH_FAILS,
+  REPORTED_WITHOUT_FAILS: jobStatuses.REPORTED_WITHOUT_FAILS,
+  KILLED: jobStatuses.KILLED,
+  FAILED: jobStatuses.FAILED,
+});
+
 export const JOB_TYPE = Object.freeze({
   FILE: "file",
   OBSERVABLE: "observable",
 });
+
 export const FILE_MIME_TYPES = Object.freeze({
   // IMPORTANT! in case you update this Object remember to update also the blackend
   WSCRIPT: "application/w-script-file",
@@ -56,6 +82,7 @@ export const FILE_MIME_TYPES = Object.freeze({
   MIXED: "multipart/mixed",
   X_SHELLSCRIPT: "text/x-shellscript",
 });
+
 export const OBSERVABLE_CLASSIFICATION = Object.freeze({
   IP: "ip",
   URL: "url",
@@ -63,72 +90,3 @@ export const OBSERVABLE_CLASSIFICATION = Object.freeze({
   HASH: "hash",
   GENERIC: "generic",
 });
-// colors
-export const JOB_TYPE_COLOR_MAP = Object.freeze({
-  file: "#ed896f",
-  observable: "#42796f",
-});
-export const TLP_COLOR_MAP = Object.freeze({
-  CLEAR: "#FFFFFF",
-  GREEN: "#33FF00",
-  AMBER: "#FFC000",
-  RED: "#FF0033",
-});
-
-export const OBSERVABLE_CLASSIFICATION_COLOR_MAP = Object.freeze({
-  ip: "#9aa66c",
-  url: "#7da7d3",
-  domain: "#8070ed",
-  hash: "#ed896f",
-  generic: "#733010",
-});
-export const TLP_DESCRIPTION_MAP = Object.freeze({
-  CLEAR: "TLP: use all analyzers",
-  GREEN: "TLP: disable analyzers that could impact privacy",
-  AMBER:
-    "TLP: disable analyzers that could impact privacy and limit access to my organization",
-  RED: "TLP: disable analyzers that could impact privacy, limit access to my organization and do not use any external service",
-});
-export const JOB_STATUS_COLOR_MAP = Object.freeze({
-  pending: "light",
-  running: "secondary",
-  analyzers_running: "secondary",
-  connectors_running: "secondary",
-  pivots_running: "secondary",
-  visualizers_running: "secondary",
-
-  analyzers_completed: "secondary",
-  connectors_completed: "secondary",
-  pivots_completed: "secondary",
-  visualizers_completed: "secondary",
-
-  killed: "gray",
-  reported_with_fails: "warning",
-  reported_without_fails: "success",
-  failed: "danger",
-});
-export const REPORT_STATUS_COLOR_MAP = Object.freeze({
-  pending: "light",
-  running: "secondary",
-  killed: "gray",
-  success: "success",
-  failed: "danger",
-});
-export const STATUS_COLORMAP = Object.freeze({
-  ...JOB_STATUS_COLOR_MAP,
-  ...REPORT_STATUS_COLOR_MAP,
-});
-export const TLP_CHOICES = Object.keys(TLP_COLOR_MAP);
-
-export const scanTypes = Object.freeze({
-  playbooks: "Playbooks",
-  analyzers_and_connectors: "Analyzers/Connectors",
-});
-
-export const HACKER_MEME_STRING =
-  "LoOk At YoU hAcKeR a PaThEtIc CrEaTuRe Of MeAt AnD bOnE";
-export const EMAIL_REGEX = /^[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,4}$/i;
-export const URL_REGEX = "(www.|http://|https://).*";
-export const UUID_REGEX =
-  /^[0-9A-F]{8}-[0-9A-F]{4}-4[0-9A-F]{3}-[89AB][0-9A-F]{3}-[0-9A-F]{12}$/i;
-export const PASSWORD_REGEX = /^(?=.*[a-zA-Z])[a-zA-Z0-9]{12,}$/i;
