+++ frontend/src/constants/miscConst.js
@@ -1,0 +1,7 @@
+export const jobResultSection = Object.freeze({
+  RAW: "raw",
+  VISUALIZER: "visualizer",
+});
+
+export const HACKER_MEME_STRING =
+  "LoOk At YoU hAcKeR a PaThEtIc CrEaTuRe Of MeAt AnD bOnE";
