+++ frontend/tests/components/utils/time.test.js
@@ -1,4 +1,4 @@
-import { processTimeMMSS } from "../../../src/utils/time";
+import { processTimeMMSS, parseScanCheckTime } from "../../../src/utils/time";
 
 describe("test time utilities", () => {
 
@@ -6,4 +6,14 @@ describe("test time utilities", () => {
     expect(processTimeMMSS(3.08)).toBe("00:03");
     expect(processTimeMMSS(0.23)).toBe("00:00");
   })
+
+  test("parseScanCheckTime - correct time: days:hours:minutes:seconds", () => {
+    const time = parseScanCheckTime("01:02:00:00");
+    expect(time).toBe(26);
+  });
+
+  test("parseScanCheckTime - not correct time: days-hours-minutes-seconds", () => {
+    const time = parseScanCheckTime("01-02-00-00");
+    expect(time).toBe(NaN);
+  });
 })
